first draft of brexiters
I measure right the position pretending in 1D or in more D, but when we measure distance only in one updated dimension.

If I pretend Brexiters position in 2+D I made it wrong and I can't fix it... Probably it needs to use goniometric functions, but I'm bad in them...

Ashwin, would you be able to fixe it?
Problem is in procedure 'set-brexit-opinion', there I measure correctly distance of agents, but in 2+D I wrongly mimick the position of Brexiters...