Create groupFormation_proportionalSampling.nlogo
Milano version with proportional sampling, i.e. this version is without Twitter agents.
NOTE! This version is not just cutted off Twitter agents! There is improvement in how we draw discussion capacity: Discusion capacity is drawn randomly and if it is less than number of cordial agents (small-world network) then it is set to 0 (no extra sampling only communication with cordial agents) if it is same or more, then the number of cordial agents is distracted from discussion capacity. I.e. capacity now stores capacity to communicate with other than cordial agents.